+++ sdk/core/azure-core/test/ut/challenge_based_authentication_policy_test.cpp
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+#include <azure/core/http/policies/policy.hpp>
+
+#include <gtest/gtest.h>
+
+TEST(ChallengeBasedAuthenticationPolicy, FirstUse)
+{
+}