+++ sdk/identity/azure-identity/src/client_secret_credential.cpp
@@ -9,22 +9,17 @@
 
 using namespace Azure::Identity;
 
-std::string const Azure::Identity::_detail::g_aadGlobalAuthority
-    = "https://login.microsoftonline.com/";
-
 ClientSecretCredential::ClientSecretCredential(
     std::string const& tenantId,
     std::string const& clientId,
     std::string const& clientSecret,
     std::string const& authorityHost,
+    bool disableTenantDiscovery,
     Azure::Core::Credentials::TokenCredentialOptions const& options)
     : m_tokenCredentialImpl(std::make_unique<_detail::TokenCredentialImpl>(options)),
-      m_isAdfs(tenantId == "adfs")
+      m_clientCredentialHelper(tenantId, authorityHost, disableTenantDiscovery)
 {
   using Azure::Core::Url;
-  m_requestUrl = Url(authorityHost);
-  m_requestUrl.AppendPath(tenantId);
-  m_requestUrl.AppendPath(m_isAdfs ? "oauth2/token" : "oauth2/v2.0/token");
 
   std::ostringstream body;
   body << "grant_type=client_credentials&client_id=" << Url::Encode(clientId)
@@ -38,7 +33,13 @@ ClientSecretCredential::ClientSecretCredential(
     std::string const& clientId,
     std::string const& clientSecret,
     ClientSecretCredentialOptions const& options)
-    : ClientSecretCredential(tenantId, clientId, clientSecret, options.AuthorityHost, options)
+    : ClientSecretCredential(
+        tenantId,
+        clientId,
+        clientSecret,
+        options.AuthorityHost,
+        options.DisableTenantDiscovery,
+        options)
 {
 }
 
@@ -52,6 +53,7 @@ ClientSecretCredential::ClientSecretCredential(
         clientId,
         clientSecret,
         _detail::g_aadGlobalAuthority,
+        _detail::ClientCredentialHelper::IsTenantDiscoveryDisabledByDefault(),
         options)
 {
 }
@@ -66,22 +68,25 @@ Azure::Core::Credentials::AccessToken ClientSecretCredential::GetToken(
     using _detail::TokenCredentialImpl;
     using Azure::Core::Http::HttpMethod;
 
+    auto const isAdfs = m_clientCredentialHelper.IsAdfs;
+
     std::ostringstream body;
     body << m_requestBody;
     {
       auto const& scopes = tokenRequestContext.Scopes;
       if (!scopes.empty())
       {
-        body << "&scope=" << TokenCredentialImpl::FormatScopes(scopes, m_isAdfs);
+        body << "&scope=" << TokenCredentialImpl::FormatScopes(scopes, isAdfs);
       }
     }
 
+    auto const requestUrl = m_clientCredentialHelper.GetRequestUrl(tokenRequestContext);
     auto request = std::make_unique<TokenCredentialImpl::TokenRequest>(
-        HttpMethod::Post, m_requestUrl, body.str());
+        HttpMethod::Post, requestUrl, body.str());
 
-    if (m_isAdfs)
+    if (isAdfs)
     {
-      request->HttpRequest.SetHeader("Host", m_requestUrl.GetHost());
+      request->HttpRequest.SetHeader("Host", requestUrl.GetHost());
     }
 
     return request;